+++ R/RcppExports.R
@@ -3285,6 +3285,10 @@ StochasticPatch___FF16__FF16_Env__ode_rates__get <- function(obj_) {
     .Call('_plant_StochasticPatch___FF16__FF16_Env__ode_rates__get', PACKAGE = 'plant', obj_)
 }
 
+StochasticPatch___FF16__FF16_Env__node_ode_size__get <- function(obj_) {
+    .Call('_plant_StochasticPatch___FF16__FF16_Env__node_ode_size__get', PACKAGE = 'plant', obj_)
+}
+
 StochasticPatch___FF16__FF16_Env__state__get <- function(obj_) {
     .Call('_plant_StochasticPatch___FF16__FF16_Env__state__get', PACKAGE = 'plant', obj_)
 }
@@ -3375,6 +3379,10 @@ StochasticPatch___TF24__TF24_Env__ode_state__get <- function(obj_) {
 
 StochasticPatch___TF24__TF24_Env__ode_rates__get <- function(obj_) {
     .Call('_plant_StochasticPatch___TF24__TF24_Env__ode_rates__get', PACKAGE = 'plant', obj_)
+}
+
+StochasticPatch___TF24__TF24_Env__node_ode_size__get <- function(obj_) {
+    .Call('_plant_StochasticPatch___TF24__TF24_Env__node_ode_size__get', PACKAGE = 'plant', obj_)
 }
 
 StochasticPatch___TF24__TF24_Env__state__get <- function(obj_) {
@@ -3469,6 +3477,10 @@ StochasticPatch___TF24f__TF24_Env__ode_rates__get <- function(obj_) {
     .Call('_plant_StochasticPatch___TF24f__TF24_Env__ode_rates__get', PACKAGE = 'plant', obj_)
 }
 
+StochasticPatch___TF24f__TF24_Env__node_ode_size__get <- function(obj_) {
+    .Call('_plant_StochasticPatch___TF24f__TF24_Env__node_ode_size__get', PACKAGE = 'plant', obj_)
+}
+
 StochasticPatch___TF24f__TF24_Env__state__get <- function(obj_) {
     .Call('_plant_StochasticPatch___TF24f__TF24_Env__state__get', PACKAGE = 'plant', obj_)
 }
@@ -3559,6 +3571,10 @@ StochasticPatch___K93__K93_Env__ode_state__get <- function(obj_) {
 
 StochasticPatch___K93__K93_Env__ode_rates__get <- function(obj_) {
     .Call('_plant_StochasticPatch___K93__K93_Env__ode_rates__get', PACKAGE = 'plant', obj_)
+}
+
+StochasticPatch___K93__K93_Env__node_ode_size__get <- function(obj_) {
+    .Call('_plant_StochasticPatch___K93__K93_Env__node_ode_size__get', PACKAGE = 'plant', obj_)
 }
 
 StochasticPatch___K93__K93_Env__state__get <- function(obj_) {